+++ includes/libraries/aihrus-framework/includes/class-aihrus-common.php
@@ -772,8 +772,6 @@ EOD;
 
 		return $value;
 	}
-
-
 }